Extract common implementation into traits
